Get static allocation working.
This should be good to use as a starting point to port to microcontroller based
project. Need to replace read and write calls with read from flash read and load
to FPGA image respectively.

Any changes in the parameters for encoder should be taken care in this program.

Downside is that parmeters here are not configurable which is fine for a
microcontroller based embedded system.

Signed-off-by: Ratheesh S <ratheeshreddy@gmail.com>
Tested-by: Ratheesh S <ratheeshreddy@gmail.com>